requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
fastapi==0.109.0
h11==0.14.0
idna==3.6
loguru==0.7.2
prompt-toolkit==3.0.43
pydantic==2.5.3
pydantic_core==2.14.6
pymongo[srv]==4.6.1
pytest==8.0.0
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.35.1
typer==0.9.0
typing_extensions==4.9.0
uvicorn==0.26.0
wcwidth==0.2.13
websockets==12.0
